requirements.txt
kafka-python
python-dotenv
happybase
torch
transformers